+++ tests/data/modules_v2/refs/inmanta_plugins/refs/__init__.py
@@ -146,7 +146,6 @@ class DictMade(Reference[str]):
         :param name: The name of the environment variable.
         """
         super().__init__()
-        breakpoint()
         self.name = name
 
     def resolve(self, logger: LoggerABC) -> str:
